requirements.txt
amqp==2.4.2
billiard==3.5.0.5
blis==0.2.4
celery==4.2.2
certifi==2019.3.9
chardet==3.0.4
Click==7.0
cymem==2.0.2
Django==2.1.2
django-background-tasks==1.1.13
django-compat==1.0.15
django-environ==0.4.1
django-extra-fields==1.0.0
django-rq==1.3.0
djangorestframework==3.9.2
idna==2.8
jsonschema==2.6.0
kombu==4.2.0
murmurhash==1.0.2
numpy==1.16.2
Pillow==6.0.0
plac==0.9.6
preshed==2.0.1
protobuf==3.7.1
psycopg2==2.7.7
psycopg2-binary==2.7.7
pytz==2019.1
redis==3.2.1
requests==2.22.0
rq==1.0
scipy==1.2.1
six==1.12.0
spacy==2.1.3
srsly==0.0.5
thinc==7.0.4
tqdm==4.32.1
urllib3==1.25.2
vine==1.3.0
wasabi==0.2.2
